requirements.txt
asgiref==3.3.1
autopep8==1.5.6
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cloudinary==1.26.0
cryptography==3.4.7
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.1.7
django-allauth==0.44.0
django-ckeditor==6.1.0
django-cloudinary-storage==0.3.0
django-countries==7.1
django-crispy-forms==1.11.2
django-debug-toolbar==3.2
django-filter==2.4.0
django-heroku==0.3.1
django-import-export==2.5.0
django-js-asset==1.2.2
django-phonenumber-field==5.0.0
django-redis==4.12.1
django-robots==4.0
et-xmlfile==1.0.1
gunicorn==20.1.0
idna==2.10
MarkupPy==1.14
oauthlib==3.1.0
odfpy==1.4.1
openpyxl==3.0.7
phonenumbers==8.12.20
Pillow==8.2.0
psycopg2==2.8.6
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.0.1
python-decouple==3.4
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
urllib3==1.26.4
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
